move portal views to routes
