+++ src/lib/app/portalsStore.ts
@@ -2,11 +2,10 @@ import {get, writable, type Writable} from 'svelte/store';
 import {setContext, getContext} from 'svelte';
 
 import {type PortalsData, type PortalData} from '$lib/portals/portal';
-import voidPortal from '$lib/portals/void/data';
 
 export interface PortalsState {
 	data: PortalsData;
-	selectedPortal: PortalData;
+	selectedPortal: PortalData | null;
 }
 
 export interface PortalsStore {
@@ -20,18 +19,15 @@ export const createPortalsStore = (initialState: PortalsState): PortalsStore => 
 	const portalsStore: PortalsStore = {
 		subscribe,
 		select: (slug: string) => {
-			if (get(store).selectedPortal.slug === slug) return;
+			if (get(store).selectedPortal?.slug === slug) return;
 			update(($portals) => ({
 				...$portals,
-				selectedPortal: findPortalBySlug($portals.data, slug),
+				selectedPortal: $portals.data.portalsBySlug.get(slug)!,
 			}));
 		},
 	};
 	return portalsStore;
 };
-
-export const findPortalBySlug = (portalsData: PortalsData, slug: string): PortalData =>
-	portalsData.portalsBySlug.get(slug) || portalsData.portalsBySlug.get(voidPortal.name)!;
 
 export const portalsContextKey = {};
 export const getPortals = (): PortalsStore => getContext(portalsContextKey);
