--- src/lib/portals/about/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import About from './About.svelte';
-</script>
-
-<About />
