--- src/lib/portals/about/view.ts
@@ -1,3 +1,0 @@
-import View from './View.svelte';
-
-export const view = View;
