--- src/lib/portals/deep-breath/View.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import DeepBreath from './DeepBreath.svelte';
-	import {getClock} from '$lib/app/clockStore';
-
-	const clock = getClock();
-</script>
-
-<DeepBreath {clock} />
