--- src/lib/portals/easings-1/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import EasingViz from './EasingViz.svelte';
-</script>
-
-<EasingViz />
