--- src/lib/portals/easings-2/View.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-	import EasingAudViz from './EasingAudViz.svelte';
-	import {getClock} from '$lib/app/clockStore';
-
-	const clock = getClock();
-</script>
-
-<EasingAudViz {clock} />
