--- src/lib/portals/hearing-test/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import HearingTest from './HearingTest.svelte';
-</script>
-
-<HearingTest />
