--- src/lib/portals/paint-freqs/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import PaintFreqs from './PaintFreqs.svelte';
-</script>
-
-<PaintFreqs />
