--- src/lib/portals/transition-designer/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import TransitionDesigner from './TransitionDesigner.svelte';
-</script>
-
-<TransitionDesigner />
