--- src/lib/portals/under-construction/View.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import UnderConstruction from './UnderConstruction.svelte';
-</script>
-
-<UnderConstruction />
