--- src/lib/portals/void/Preview.svelte
@@ -1,1 +1,0 @@
-{'{'} void {'}'}
