--- src/lib/portals/void/data.ts
@@ -1,11 +1,0 @@
-import {type PortalData} from '$lib/portals/portal';
-import Preview from './Preview.svelte';
-
-const data: PortalData = {
-	name: 'void',
-	slug: 'void',
-	showHomeButton: true,
-	Preview,
-};
-
-export default data;
