+++ src/routes/paint-freqs.svelte
